Update Integer to Long

+++ lib/src/main/java/org/eclipse/tsp/java/client/api/experiment/dto/UpdateExperimentRequestDto.java
@@ -32,7 +32,7 @@ public class UpdateExperimentRequestDto {
 		@NonNull
 		private UpdateTraceAction action = UpdateTraceAction.ADD;
 
-		private Integer timeOffset;
+		private Long timeOffset;
 	}
 
 	public enum UpdateTraceAction {
